requirements.txt
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
pyyaml>=5.4.1
python-dotenv>=0.19.0

sqlalchemy